fea: modification and make otp app.
1.modified url by adding method name at comment, views by adding comment goal or action, model bychanging statements and move otp model from accounts app to otp app, settings by rest framework settings and move jwt settings from settings to jwt_settings file.

2.make otp app for place all functionality of otp generate and save and send.
