+++ jwt_authentication/accounts/api/urls.py
@@ -6,15 +6,16 @@ from rest_framework_simplejwt.views import (
     TokenRefreshView,
 )
 
+
 urlpatterns = [
-    # localhost:8000/
+    # GET: localhost:8000/
     path("", getRoutes),
-    # localhost:8000/api/token/
+    # POST: localhost:8000/api/token/
     path(
         route="token/", view=MyTokenObtainPairView.as_view(), name="token_obtain_pair"
     ),
     # localhost:8000/api/token/refresh/
     path(route="token/refresh/", view=TokenRefreshView.as_view(), name="token_refresh"),
-    # localhost:8000/api/home/
+    # GET: localhost:8000/api/home/
     path(route="home/", view=home_view.as_view(), name="home"),
 ]
