+++ jwt_authentication/accounts/api/views.py
@@ -9,11 +9,21 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
 
+@api_view(["GET"])
+def getRoutes(request):
+
+    routes = ["/api/token/", "/api/token/refresh/", "/api/home/"]
+
+    return Response(routes)
+
+
 class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
     """This class is a custom serializer for obtaining authentication tokens"""
 
     @classmethod
     def get_token(cls, user):
+        """Used to get token and set user data"""
+
         token = super().get_token(user)
         token["email"] = user.email
         token["first_name"] = user.first_name
@@ -21,10 +31,12 @@ class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
         token["current_date"] = datetime.now().strftime("%Y:%m:%d")
         current_time = datetime.now().strftime("%I:%M:%p")
         token["current_time"] = current_time
+
         if "AM" in current_time:
             token["day_status"] = "Day"
         else:
             token["day_status"] = "Night"
+
         token["location"] = "Dhaka"
 
         return token
@@ -36,23 +48,21 @@ class MyTokenObtainPairView(TokenObtainPairView):
     serializer_class = CustomTokenObtainPairSerializer
 
     def post(self, request, *args, **kwargs):
+        """Used to return access token and refresh token"""
+
         response = super().post(request, *args, **kwargs)
         access_token = str(response.data["access"])
         refresh_token = str(response.data["refresh"])
+
         token_data = {
             "access": access_token,
             "refresh": refresh_token,
         }
+
         token = access_token
         request.session["token"] = token
+
         return Response(token_data)
-
-
-@api_view(["GET"])
-def getRoutes(request):
-
-    routes = ["/api/token/", "/api/token/refresh/", "/api/home/"]
-    return Response(routes)
 
 
 class home_view(APIView):
@@ -61,4 +71,6 @@ class home_view(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request, *args, **kwargs):
+        """Used to return success message"""
+
         return Response({"message": "Welcome to the home view!"})
