+++ jwt_authentication/accounts/models.py
@@ -7,22 +7,34 @@ from django.db import models
 
 
 class CustomUserManager(BaseUserManager):
+    """used to help creating user instant"""
+
     def create_user(self, email, password=None, **extra_fields):
+        """Used to create  user"""
+
         if not email:
             raise ValueError("The Email field must be set")
+
         email = self.normalize_email(email)
         user = self.model(email=email, **extra_fields)
         user.set_password(password)
         user.save(using=self._db)
+
         return user
 
     def create_superuser(self, email, password=None, **extra_fields):
+        """Used to create super user"""
+
         extra_fields.setdefault("is_staff", True)
+        extra_fields.setdefault("is_active", True)
         extra_fields.setdefault("is_superuser", True)
+
         return self.create_user(email, password, **extra_fields)
 
 
 class CustomUser(AbstractBaseUser, PermissionsMixin):
+    """Custom user model"""
+
     email = models.EmailField(unique=True)
     first_name = models.CharField(max_length=100)
     last_name = models.CharField(max_length=100)
@@ -32,21 +44,20 @@ class CustomUser(AbstractBaseUser, PermissionsMixin):
 
     objects = CustomUserManager()
 
-    groups = None
-    user_permissions = None
-
     USERNAME_FIELD = "email"
     REQUIRED_FIELDS = ["first_name", "last_name"]
 
+    def get_full_name(self):
+        """Used to get user full name"""
+
+        return self.first_name + self.last_name
+
+    def get_short_name(self):
+        """used to get short name"""
+
+        return self.first_name
+
     def __str__(self):
-        return self.email
+        """Return email"""
 
-
-class OTP(models.Model):
-    email = models.EmailField(unique=True)
-    otp = models.CharField(max_length=6)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
         return self.email
