+++ jwt_authentication/accounts/urls.py
@@ -6,16 +6,17 @@ from .views import (
     ForgottenPasswordResetView,
 )
 
+
 urlpatterns = [
-    # localhost:8000/api-auth/register/
+    # POST: localhost:8000/api-auth/register/
     path(
         route="register/", view=UserRegistrationView.as_view(), name="user_registration"
     ),
-    # localhost:8000/api-auth/activate/
+    # PATCH: localhost:8000/api-auth/activate/
     path(route="activate/", view=UserActivationView.as_view(), name="user_activation"),
-    # localhost:8000/api-auth/reset/
+    # PATCH: localhost:8000/api-auth/reset/
     path(route="reset/", view=UserPasswordResetView.as_view(), name="change_password"),
-    # localhost:8000/api-auth/recover/
+    # PATCH: localhost:8000/api-auth/recover/
     path(
         route="recover/",
         view=ForgottenPasswordResetView.as_view(),
