+++ jwt_authentication/otp/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from .models import OTP
+
+
+class AuthenticationAdmin(admin.ModelAdmin):
+    readonly_fields = ("created_at", "updated_at")
+
+
+admin.site.register(OTP)
