+++ jwt_authentication/otp/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class OtpConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "otp"
